+++ src/main/java/at/hannibal2/skyhanni/config/features/Garden.java
@@ -22,4 +22,33 @@ public class Garden {
     @ConfigEditorButton(runnableId = "skyMartCopperPrice", buttonText = "Edit")
     @ConfigAccordionId(id = 0)
     public Position skyMartCopperPricePos = new Position(44, -108, false, true);
+
+    @Expose
+    @ConfigOption(name = "Visitor Helper", desc = "")
+    @ConfigEditorAccordion(id = 1)
+    public boolean visitorHelper = false;
+
+    @Expose
+    @ConfigOption(name = "Visitor Display", desc = "Show all items needed for the visitors.")
+    @ConfigEditorBoolean
+    @ConfigAccordionId(id = 1)
+    public boolean visitorHelperDisplay = true;
+
+    @Expose
+    @ConfigOption(name = "Visitor Helper Position", desc = "")
+    @ConfigEditorButton(runnableId = "visitorHelper", buttonText = "Edit")
+    @ConfigAccordionId(id = 1)
+    public Position visitorHelperPos = new Position(0, 0, false, true);
+
+    @Expose
+    @ConfigOption(name = "Highlight Ready", desc = "Highlight the visitor when the required items are in the inventory.")
+    @ConfigEditorBoolean
+    @ConfigAccordionId(id = 1)
+    public boolean visitorHelperHighlightReady = true;
+
+    @Expose
+    @ConfigOption(name = "Show Price", desc = "Show the bazaar price of the items required for the visitors.")
+    @ConfigEditorBoolean
+    @ConfigAccordionId(id = 1)
+    public boolean visitorHelperShowPrice = true;
 }